--- Ground Truth Merkle.py
@@ -1,1 +1,0 @@
-
