--- hash.py
@@ -1,58 +1,0 @@
-import numpy as np
-import string
-import hashlib
-import poseidon
-
-prime_254 = poseidon.parameters.prime_254
-prime_255 = poseidon.parameters.prime_255
-
-def SHA(s: string) -> string:
-    return hashlib.sha256(s.encode()).hexdigest()
-
-def toDigit(s: string) -> int:
-    return int(s, 16)
-
-class PoseidonHash(poseidon.Poseidon):
-    def __init__(self, prime=prime_254, security_level=128, input_rate=3, t=4, alpha=5):
-        """
-        # TODO - add parameters description
-        """
-        super().__init__(prime, security_level, alpha, input_rate, t)
-    
-    def __str__(self) -> str:
-        return "PoseidonHash"
-
-
-# # poseidonperm_x5_255_3
-# Input:
-# ['0x0000000000000000000000000000000000000000000000000000000000000000', '0x0000000000000000000000000000000000000000000000000000000000000001', '0x0000000000000000000000000000000000000000000000000000000000000002']
-# Output:
-# ['0x28ce19420fc246a05553ad1e8c98f5c9d67166be2c18e9e4cb4b4e317dd2a78a', '0x51f3e312c95343a896cfd8945ea82ba956c1118ce9b9859b6ea56637b4b1ddc4', '0x3b2b69139b235626a0bfb56c9527ae66a7bf486ad8c11c14d1da0c69bbe0f79a']
-
-# # poseidonperm_x5_255_5
-# Input:
-# ['0x0000000000000000000000000000000000000000000000000000000000000000', '0x0000000000000000000000000000000000000000000000000000000000000001', '0x0000000000000000000000000000000000000000000000000000000000000002', '0x0000000000000000000000000000000000000000000000000000000000000003', '0x0000000000000000000000000000000000000000000000000000000000000004']
-# Output:
-# ['0x2a918b9c9f9bd7bb509331c81e297b5707f6fc7393dcee1b13901a0b22202e18', '0x65ebf8671739eeb11fb217f2d5c5bf4a0c3f210e3f3cd3b08b5db75675d797f7', '0x2cc176fc26bc70737a696a9dfd1b636ce360ee76926d182390cdb7459cf585ce', '0x4dc4e29d283afd2a491fe6aef122b9a968e74eff05341f3cc23fda1781dcb566', '0x03ff622da276830b9451b88b85e6184fd6ae15c8ab3ee25a5667be8592cce3b1']
-
-# # poseidonperm_x5_254_3
-# Input:
-# ['0x0000000000000000000000000000000000000000000000000000000000000000', '0x0000000000000000000000000000000000000000000000000000000000000001', '0x0000000000000000000000000000000000000000000000000000000000000002']
-# Output:
-# ['0x115cc0f5e7d690413df64c6b9662e9cf2a3617f2743245519e19607a4417189a', '0x0fca49b798923ab0239de1c9e7a4a9a2210312b6a2f616d18b5a87f9b628ae29', '0x0e7ae82e40091e63cbd4f16a6d16310b3729d4b6e138fcf54110e2867045a30c']
-
-# # poseidonperm_x5_254_5
-# Input:
-# ['0x0000000000000000000000000000000000000000000000000000000000000000', '0x0000000000000000000000000000000000000000000000000000000000000001', '0x0000000000000000000000000000000000000000000000000000000000000002', '0x0000000000000000000000000000000000000000000000000000000000000003', '0x0000000000000000000000000000000000000000000000000000000000000004']
-# Output:
-# ['0x299c867db6c1fdd79dcefa40e4510b9837e60ebb1ce0663dbaa525df65250465', '0x1148aaef609aa338b27dafd89bb98862d8bb2b429aceac47d86206154ffe053d', '0x24febb87fed7462e23f6665ff9a0111f4044c38ee1672c1ac6b0637d34f24907', '0x0eb08f6d809668a981c186beaf6110060707059576406b248e5d9cf6e78b3d3e', '0x07748bc6877c9b82c8b98666ee9d0626ec7f5be4205f79ee8528ef1c4a376fc7']
-
-# # starkadperm_x5_256_3
-# Input:
-# ['0x0000000000000000000000000000000000000000000000000000000000000000', '0x0000000000000000000000000000000000000000000000000000000000000001', '0x0000000000000000000000000000000000000000000000000000000000000002']
-# Output:
-# ['0xd4a965f154ff8634896b30da71a3b5dfdbf8f87bba6dd053c7dc32b289464a50', '0x21edf7423857b5fd375ac0808409b50e7b92c580fe0b8415bb0a8cf352995118', '0x73b5e0fb2390fd139c9fa8a3e3a2c2349e504b02f80eef19e434070fa4ce147d']
-# Input (concat):
-# 0x000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000002
-# Output (concat):
-# 0xd4a965f154ff8634896b30da71a3b5dfdbf8f87bba6dd053c7dc32b289464a5021edf7423857b5fd375ac0808409b50e7b92c580fe0b8415bb0a8cf35299511873b5e0fb2390fd139c9fa8a3e3a2c2349e504b02f80eef19e434070fa4ce147d
