--- poseidon-hash/CHANGELOG.md
@@ -1,5 +1,0 @@
-# CHANGELOG
-
-## Version 0.1.0 07/18/2022
-1. First major release of the code, includes implementation of hash function and its optimization.
-2. Add generator of round constants, round numbers and MDS matrix with correspond tests.
