--- poseidon-hash/MANIFEST.in
@@ -1,1 +1,0 @@
-global-include *.txt *.py *.md
