--- poseidon-hash/poseidon/__init__.py
@@ -1,2 +1,0 @@
-from .hash import Poseidon, OptimizedPoseidon, HashType
-from .parameters import *
