--- poseidon-hash/pyproject.toml
@@ -1,3 +1,0 @@
-[build-system]
-requires = ["setuptools>=42.0", "wheel"]
-build-backend = "setuptools.build_meta"
