--- poseidon-hash/requirements.txt
@@ -1,5 +1,0 @@
-galois==0.0.30
-numpy==1.22.4
-py_ecc==6.0.0
-pytest==7.1.3
-setuptools==65.6.3
