+++ py_cc/__init__.py
@@ -1,0 +1,3 @@
+VERSION = '0.0.1'
+from py_cc import elliptic_curves
+from py_cc import hashes
